fix DeleteDialog
