+++ src/index.ts
@@ -15,7 +15,7 @@ export { default as BoxContent } from './components/Box/BoxContent'
 export { default as BoxWrap } from './components/Box/BoxWrap'
 export { ButtonWrap } from './components/Button/ButtonWrap'
 export { CustomButton } from './components/Button/CustomButton'
-export { default as DeleteDialog } from './components/DeleteDialog'
+export { default as ConfirmDialog } from './components/ConfirmDialog'
 export { CustomLabel } from './components/Form/CustomLabel'
 export { RadioButton } from './components/Form/RadioButton'
 export { default as MainTitleSet } from './components/Text/MainTitleSet'
